requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain_community
langchain_groq
chromadb
python-dotenv
transformers
torch
numpy
